fixed editor premake file did not define debug in debug mode

+++ Editor/src/Utilities/FontsManager/FontManager.h
@@ -9,7 +9,7 @@
 struct FontInfo {
 	std::string name;
 	std::filesystem::path path;
-	uint32_t size;
+	float size;
 };
